Fixed deletion code to not trust post request.
This change makes the server get the project data
from the database and checks it's customer id to
see if the logged in user's id is the same to ensure
another non-admin user can't delete the project. Also,
I made sure to convert the IDs to strings so they could
be properly compared.